# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/reports": {
      "get": {
        "operationId": "ReportController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "profile_id",
            "required": false,
            "in": "query",
            "description": "Filter by profile id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "description": "Filter by kind",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List reports with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reports"
        ]
      },
      "post": {
        "operationId": "ReportController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/reports/{id}": {
      "get": {
        "operationId": "ReportController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reports"
        ]
      },
      "patch": {
        "operationId": "ReportController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reports"
        ]
      },
      "delete": {
        "operationId": "ReportController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/tasks": {
      "get": {
        "operationId": "TasksController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "default": 50,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List tasks with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasksListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "operationId": "TasksController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Trade API",
    "description": "Auto-generated OpenAPI documentation for Trade backend",
    "version": "1.0.0",
    "contact": {
      "name": "API",
      "url": "",
      "email": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "ReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Report UUID"
          },
          "profile_id": {
            "type": "number",
            "description": "Profile id this report belongs to",
            "example": 1
          },
          "job_id": {
            "type": "object",
            "nullable": true,
            "description": "Job id that produced this report"
          },
          "content": {
            "type": "object",
            "nullable": true,
            "description": "Main LLM text (narrative/outlook)"
          },
          "confidence": {
            "type": "object",
            "nullable": true,
            "description": "Confidence score"
          },
          "llmModel": {
            "type": "object",
            "nullable": true,
            "description": "LLM model used"
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp (ISO 8601)",
            "example": "2025-01-01T12:00:00.000Z"
          },
          "kind": {
            "type": "string",
            "description": "Kind/category of report",
            "example": "default"
          },
          "relevance": {
            "type": "object",
            "nullable": true,
            "description": "Relevance to the profile/topic (0..1)"
          },
          "tokens_in": {
            "type": "object",
            "nullable": true,
            "description": "Input tokens used by the LLM"
          },
          "tokens_out": {
            "type": "object",
            "nullable": true,
            "description": "Output tokens generated by the LLM"
          },
          "cost": {
            "type": "object",
            "nullable": true,
            "description": "Approximate processing cost (USD)"
          },
          "summary_bullets": {
            "nullable": true,
            "description": "Short bullet-point summary",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "signals": {
            "nullable": true,
            "description": "Key signals or insights",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "risks": {
            "nullable": true,
            "description": "Risks or caveats",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sentiment_label": {
            "type": "string",
            "enum": [
              "bearish",
              "neutral",
              "bullish"
            ],
            "nullable": true,
            "description": "Overall sentiment label"
          },
          "sentiment_score": {
            "type": "number",
            "nullable": true,
            "description": "Numeric sentiment score"
          },
          "outlook": {
            "type": "string",
            "nullable": true,
            "description": "Forward-looking outlook text"
          },
          "support": {
            "description": "Citations with documentId and quote",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "profile_id",
          "job_id",
          "content",
          "confidence",
          "llmModel",
          "created_at",
          "kind",
          "relevance",
          "tokens_in",
          "tokens_out",
          "cost",
          "summary_bullets",
          "signals",
          "risks",
          "sentiment_label",
          "sentiment_score",
          "outlook",
          "support"
        ]
      },
      "ReportsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Returned reports",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of reports matching the query",
            "example": 123
          },
          "limit": {
            "type": "number",
            "description": "Page size (limit) applied to the query",
            "example": 50
          },
          "offset": {
            "type": "number",
            "description": "Offset applied to the query",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "ReportCreateDto": {
        "type": "object",
        "properties": {
          "profile_id": {
            "type": "number",
            "description": "Profile id to attach report to"
          },
          "content": {
            "type": "object",
            "description": "Main LLM text (narrative/outlook)"
          },
          "llmModel": {
            "type": "object",
            "description": "LLM model used"
          },
          "job_id": {
            "type": "object",
            "description": "Job id that produced this report"
          },
          "confidence": {
            "type": "object",
            "description": "Confidence score"
          },
          "kind": {
            "type": "string",
            "description": "Kind/category of report",
            "default": "default"
          },
          "relevance": {
            "type": "object",
            "description": "How relevant this report is to the profile/topic (0..1)",
            "example": 0.87
          },
          "tokens_in": {
            "type": "object",
            "description": "Number of input tokens used by the LLM",
            "example": 1234
          },
          "tokens_out": {
            "type": "object",
            "description": "Number of output tokens generated by the LLM",
            "example": 512
          },
          "cost": {
            "type": "object",
            "description": "Approximate processing cost (USD)",
            "example": 0.023
          },
          "summary_bullets": {
            "description": "Short bullet-point summary generated by the model",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "signals": {
            "description": "Key signals or insights identified",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "risks": {
            "description": "Risks or caveats mentioned in the report",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sentiment_label": {
            "type": "string",
            "enum": [
              "bearish",
              "neutral",
              "bullish"
            ],
            "description": "Overall sentiment label"
          },
          "sentiment_score": {
            "type": "number",
            "description": "Numeric sentiment score (e.g., -1..1 or 0..1 depending on model)"
          },
          "outlook": {
            "type": "string",
            "description": "Forward-looking outlook text"
          },
          "support": {
            "description": "Citations with documentId and quote",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "profile_id"
        ]
      },
      "ReportUpdateDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "object",
            "description": "Main LLM text (narrative/outlook)"
          },
          "llmModel": {
            "type": "object",
            "description": "LLM model used"
          },
          "job_id": {
            "type": "object",
            "description": "Job id that produced this report"
          },
          "confidence": {
            "type": "object",
            "description": "Confidence score"
          },
          "kind": {
            "type": "string",
            "description": "Kind/category of report"
          },
          "relevance": {
            "type": "object",
            "description": "How relevant this report is to the profile/topic (0..1)"
          },
          "tokens_in": {
            "type": "object",
            "description": "Number of input tokens used by the LLM"
          },
          "tokens_out": {
            "type": "object",
            "description": "Number of output tokens generated by the LLM"
          },
          "cost": {
            "type": "object",
            "description": "Approximate processing cost (USD)"
          },
          "summary_bullets": {
            "description": "Short bullet-point summary generated by the model",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "signals": {
            "description": "Key signals or insights identified",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "risks": {
            "description": "Risks or caveats mentioned in the report",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sentiment_label": {
            "type": "string",
            "enum": [
              "bearish",
              "neutral",
              "bullish"
            ],
            "description": "Overall sentiment label"
          },
          "sentiment_score": {
            "type": "number",
            "description": "Numeric sentiment score (e.g., -1..1 or 0..1 depending on model)"
          },
          "outlook": {
            "type": "string",
            "description": "Forward-looking outlook text"
          },
          "support": {
            "description": "Citations with documentId and quote",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Task numeric identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Task name",
            "example": "Summarize article"
          },
          "description": {
            "type": "object",
            "description": "Optional task description",
            "example": "Generate concise summary bullets"
          },
          "prompt": {
            "type": "string",
            "description": "Prompt/template used for the task"
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp (ISO 8601)",
            "example": "2025-01-01T12:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "description": "Last update timestamp (ISO 8601)",
            "example": "2025-01-02T08:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "prompt",
          "created_at",
          "updated_at"
        ]
      },
      "TasksListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Returned tasks",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of tasks matching the query",
            "example": 123
          },
          "limit": {
            "type": "number",
            "description": "Page size (limit) applied to the query",
            "example": 50
          },
          "offset": {
            "type": "number",
            "description": "Offset applied to the query",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "TaskCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Task name",
            "example": "Summarize article"
          },
          "description": {
            "type": "object",
            "description": "Optional task description"
          },
          "prompt": {
            "type": "string",
            "description": "Prompt/template used for the task"
          }
        },
        "required": [
          "name",
          "prompt"
        ]
      },
      "TaskUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Task name"
          },
          "description": {
            "type": "object",
            "description": "Optional task description"
          },
          "prompt": {
            "type": "string",
            "description": "Prompt/template used for the task"
          }
        }
      },
...
    }
  }
}